+++ example/src/App.vue
@@ -1,9 +1,0 @@
-<template>
-  <div> hello world </div>
-</template>
-
-<style>
-body {
-  background: white;
-}
-</style>