+++ packages/core/tsup.config.ts
@@ -1,0 +1,10 @@
+import type { Options } from 'tsup'
+
+export const tsup: Options = {
+  splitting: false,
+  sourcemap: false,
+  clean: true,
+  entryPoints: ['src/index.ts'],
+  format: ['cjs', 'esm'],
+  dts: true,
+}